+ L'ancien modèle est maintenant entièrement remplacé par le nouveau

+++ app/Http/Controllers/AdresseController.php
@@ -18,7 +18,15 @@ class AdresseController extends Controller{
     function ajout(Request $request){
         $data = $request->post();
 
-        $utilisateur = $this->utilisateurService->getAuthenticatedUser($request);
+        try{
+            $utilisateur = $this->utilisateurService->getAuthenticatedUser($request);
+            if($utilisateur == null){
+                throw Exceptions::createError(525);
+            }
+        }catch(\Exception $e){
+            $e = Exceptions::createError(525);
+            return response()->json($e->getMessage(),$e->httpCode);
+        }
         $ville = $this->villeService->findById(intval($data['Ville']['ID']));
         $adresse = $this->adresseService->add(intval($data["Numéro"]), $data["Nom de rue"], $ville, $utilisateur);
 
@@ -27,12 +35,17 @@ class AdresseController extends Controller{
 
     function supprimer(Request $request){
         $data = $request->post();
-        $this->utilisateurService->getAuthenticatedUser($request); // Sécurité
+        $user = $this->utilisateurService->getAuthenticatedUser($request); // Sécurité
+
+        if($user == null){
+            $e = Exceptions::createError(525);
+            return response()->json($e->getMessage(),$e->httpCode);
+        }
         try {
             $this->adresseService->delete($this->adresseService->findById($data["ID"]));
         } catch(\Exception $e){
             if($e->getCode() == 23000){
-                $e = Exceptions::createError(523);
+                $e = Exceptions::createError(525);
             }
             return response($e->getMessage(),$e->getCode());
         }
